refactored Initializer
