+++ src/test/scala/com/databricks/labs/overwatch/pipeline/InitializerFunctionsTest.scala
@@ -6,10 +6,6 @@ import org.scalatest.{Ignore, color}
 import org.scalatest.funspec.AnyFunSpec
 
 class InitializerFunctionsTest extends AnyFunSpec with DataFrameComparer with SparkSessionTestWrapper {
-  val conf = new Config
-  conf.setOrganizationId("demo")
-  conf.setDeploymentType("default")
-  val Initializer = new Initializer(conf, disableValidations = false, isSnap = false, initDB = true)
 
   describe("Tests for InitializerFunctions.loadLocalResource") {
     it("should throw exception for non-existing resource") {
